+++ src/space_saving.cc
@@ -1,18 +1,26 @@
 //
-// Created by Nikolai Karpov on 2019-02-24.
+// Created by nk-karpov on 2019-05-30.
 //
-#include "streamingcc_include/MisraGries.h"
+
+#include "streamingcc_include/space_saving.h"
+
+#include <cassert>
 
 namespace streamingcc {
 
 namespace integer {
 
-void MisraGriesInt::ProcessItem(const uint32_t item, const double weight) {
+void SpaceSavingInt::ProcessItem(const uint32_t item, const double weight) {
   assert(weight >= 0.);
   auto pred = [&](const counter &x) { return x.first == item; };
   auto ind = std::distance(counters_.begin(), std::find_if(counters_.begin(), counters_.end(), pred));
+  auto min_value = !counters_.empty() ? counters_.back().second : 0.;
   if (ind == counters_.size()) {
-    counters_.emplace_back(counter(item, weight));
+    if (counters_.size() == capacity_) {
+      counters_.emplace_back(counter(item, min_value + weight));
+    } else {
+      counters_.emplace_back(counter(item, weight));
+    }
   } else {
     counters_[ind].second += weight;
   }
@@ -20,25 +28,21 @@ void MisraGriesInt::ProcessItem(const uint32_t item, const double weight) {
             counters_.end(),
             [](const counter &a, const counter &b) { return a.second > b.second; });
   if (counters_.size() > capacity_) {
-    auto val = counters_.back().second;
     counters_.pop_back();
-    for (auto &x : counters_) {
-      x.second -= val;
-    }
   }
 }
 
-double MisraGriesInt::GetEstimation(const uint32_t item) const {
+double SpaceSavingInt::GetEstimation(const uint32_t item) const {
   auto pred = [&](const counter &x) { return x.first == item; };
   auto it = std::find_if(counters_.begin(), counters_.end(), pred);
+  auto min_value = !counters_.empty() ? counters_.back().second : 0;
   if (it != counters_.end()) {
     return it->second;
   } else {
-    return 0.;
+    return min_value;
   }
 }
 
 } //namespace integer
 
 } //namespace streamingcc
-
