+++ src/streamingcc_include/misra_gries.h
@@ -2,22 +2,25 @@
 // Created by nk-karpov on 2019-05-30.
 //
 
-#ifndef STREAMINGCC_SPACESAVING_H
-#define STREAMINGCC_SPACESAVING_H
+#ifndef STREAMINGCC_MISRA_GRIES_H
+#define STREAMINGCC_MISRA_GRIES_H
 
 #include "../streamingcc_include/streaming_algorithm.h"
+
+#include <algorithm>
+#include <vector>
 
 namespace streamingcc {
 
 namespace integer {
 
-class SpaceSavingInt : public StreamingAlgorithmWeightedInt {
+class MisraGriesInt : public StreamingAlgorithmWeightedInt {
  public:
-  explicit SpaceSavingInt(const size_t capacity)
+  explicit MisraGriesInt(const size_t capacity)
       : capacity_(capacity) {}
   void ProcessItem(const uint32_t item, const double weight) override;
   double GetEstimation(const uint32_t item) const;
-  ~SpaceSavingInt() override = default;
+  ~MisraGriesInt() override = default;
  private:
   using counter = std::pair<uint32_t, double>;
   std::vector<counter> counters_;
@@ -28,4 +31,4 @@ class SpaceSavingInt : public StreamingAlgorithmWeightedInt {
 
 }  // namespace streamingcc
 
-#endif //STREAMINGCC_SPACESAVING_H
+#endif //STREAMINGCC_MISRA_GRIES_H
