+++ src/streamingcc_include/space_saving.h
@@ -2,25 +2,22 @@
 // Created by nk-karpov on 2019-05-30.
 //
 
-#ifndef STREAMINGCC_MISRAGRIES_H
-#define STREAMINGCC_MISRAGRIES_H
+#ifndef STREAMINGCC_SPACE_SAVING_H
+#define STREAMINGCC_SPACE_SAVING_H
 
 #include "../streamingcc_include/streaming_algorithm.h"
-
-#include <algorithm>
-#include <vector>
 
 namespace streamingcc {
 
 namespace integer {
 
-class MisraGriesInt : public StreamingAlgorithmWeightedInt {
+class SpaceSavingInt : public StreamingAlgorithmWeightedInt {
  public:
-  explicit MisraGriesInt(const size_t capacity)
+  explicit SpaceSavingInt(const size_t capacity)
       : capacity_(capacity) {}
   void ProcessItem(const uint32_t item, const double weight) override;
   double GetEstimation(const uint32_t item) const;
-  ~MisraGriesInt() override = default;
+  ~SpaceSavingInt() override = default;
  private:
   using counter = std::pair<uint32_t, double>;
   std::vector<counter> counters_;
@@ -31,4 +28,4 @@ class MisraGriesInt : public StreamingAlgorithmWeightedInt {
 
 }  // namespace streamingcc
 
-#endif //STREAMINGCC_MISRAGRIES_H
+#endif //STREAMINGCC_SPACE_SAVING_H
